java 8 refactoring
