+++ pom.xml
@@ -21,6 +21,10 @@
       <organization>Cognitect</organization>
       <organizationUrl>http://cognitect.com</organizationUrl>
     </developer>
+        <developer>
+      <name>Slim Ouertani</name>
+      <email>ouertani@gmail.com</email>
+    </developer>
   </developers>
   <scm>
     <connection>scm:git:git@github.com:cognitect/transit-java.git</connection>
@@ -42,7 +46,7 @@
         <artifactId>maven-compiler-plugin</artifactId>
         <version>3.1</version>
         <configuration>
-          <source>1.6</source>
+          <source>1.8</source>
 	  <target>1.8</target>
 	  <compilerArgument>-Xlint:unchecked,rawtypes</compilerArgument>
         </configuration>
