--- src/main/java/com/cognitect/transducers/ATransducer.java
@@ -1,42 +1,0 @@
-// Copyright 2014 Cognitect. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS-IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package com.cognitect.transducers;
-
-/**
- * Abstract base class for implementing a transducer that implements
- * comp. Derived classes must implement apply.
- * @param <B> The type of data processed by an input process
- * @param <C> The type of data processed by the transduced process
- **/
-public abstract class ATransducer<B, C> implements ITransducer<B, C> {
-
-    /**
-     * Composes a transducer with another transducer, yielding
-     * a new transducer.
-     * @param right the transducer to compose with this transducer
-     * @param <A> the type of input processed by the reducing function
-     *           the composed transducer returns when applied
-     * @return A new composite transducer
-     */
-    @Override
-    public <A> ITransducer<A, C> comp(final ITransducer<A, B> right) {
-        return new ATransducer<A, C>() {
-            @Override
-            public <R> IReducingFunction<R, C> apply(IReducingFunction<R, ? super A> xf) {
-                return ATransducer.this.apply(right.apply(xf));
-            }
-        };
-    }
-}
