--- src/main/java/com/cognitect/transducers/Function.java
@@ -1,29 +1,0 @@
-// Copyright 2014 Cognitect. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS-IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package com.cognitect.transducers;
-
-/**
- * A function of one argument
- * @param <T> type of argument
- * @param <R> return type
- */
-public interface Function<T, R> {
-    /**
-     * Invokes the function
-     * @param t argument
-     * @return return value
-     */
-    R apply(T t);
-}
