+++ src/main/java/com/cognitect/transducers/IStepFunction.java
@@ -7,6 +7,7 @@ import java.util.concurrent.atomic.AtomicBoolean;
  * @param <R> Type of first argument and return value
  * @param <T> Type of input to reduce
  */
+@FunctionalInterface
 public interface IStepFunction<R, T> {
     /**
      * Applies the reducing function to the current result and
